chore: login, register layout
