+++ shared/src/ui/index.ts
@@ -1,4 +1,4 @@
 export * from "./provider/customThemeProvider";
 export * from "./buttons/segmentButton"
-export * from "./sidebar/sidebar"
-export * from "./header/header"
+export * from "./layout/defaultLayout"
+export * from "./layout/authLayout"
